Move LoggingRoute class to utils subfolder for use across files
Pervious implementation within app.py was causing issues with circular imports
