+++ app/app.py
@@ -1,7 +1,5 @@
-from fastapi import FastAPI, Request, Response
-from fastapi.routing import APIRoute
-from http import HTTPStatus
-from .api import master_router
+from fastapi import FastAPI
+from .api import master_router, LoggingRoute
 from contextlib import asynccontextmanager
 from config.logging_config import fastapi_logging, healthcheck_logging
 import logging
@@ -12,38 +10,6 @@ fastapi_logging()
 healthcheck_logging()
 app_logger = logging.getLogger('fastapi_logger')
 health_check_logger = logging.getLogger('health_check_logger')
-
-
-
-
-class LoggingRoute(APIRoute):
-    """
-    Class Overview:
-    Custom middleware class to log HTTP response details for each request processed by the FastAPI application.
-
-    Function Logic:
-    1. Overrides the `get_route_handler` method from `APIRoute`.
-    2. Wraps the original route handler with a custom handler that intercepts the response.
-    3. Logs the HTTP response details:
-        - If the request URL contains "/health", it uses the health check logger.
-        - Otherwise, it uses the default app logger.
-    4. Retrieves the status phrase for the response status code, defaulting to "Unknown" if the code is not standard.
-
-    Returns:
-    - The original response object after logging details.
-    """
-    def get_route_handler(self):
-
-        original_route_handler = super().get_route_handler()
-        async def custom_route_handler(request: Request):
-
-            response: Response = await original_route_handler(request)
-            logger = app_logger if "/health" not in str(request.url) else health_check_logger
-            status_phrase = HTTPStatus(response.status_code).phrase if response.status_code in HTTPStatus._value2member_map_ else "Unknown"
-            logger.debug(f"HTTP Response: {response.status_code} {status_phrase}")
-            return response
-        
-        return custom_route_handler
